Move recurrence logic to service layer
