+++ src/agendable/services/__init__.py
@@ -6,6 +6,7 @@ from agendable.services.oidc_service import (
     resolve_oidc_link_resolution,
     resolve_oidc_login_resolution,
 )
+from agendable.services.reminder_delivery_service import run_due_reminders
 from agendable.services.series_service import create_series_with_occurrences
 
 __all__ = [
@@ -16,4 +17,5 @@ __all__ = [
     "provision_user_for_oidc",
     "resolve_oidc_link_resolution",
     "resolve_oidc_login_resolution",
+    "run_due_reminders",
 ]
